requirements.txt
basyx-python-sdk==1.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
lxml==4.9.4
pip==23.3.2
pyecma376-2==1.0.0
python-dateutil==2.8.2
requests==2.31.0
setuptools==69.0.3
six==1.16.0
urllib3==1.26.18
wheel==0.42.0
